Reorganizing python structure and setting up cabled PHSEN harvester
