+++ instruments/python/uncabled/examples/ctdbp/telemetered_ctdbp.py
@@ -1,10 +1,13 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 import os
+import yaml
 
 from instruments.python.common import list_deployments, deployment_dates, get_vocabulary, m2m_request, m2m_collect, \
-    update_dataset, CONFIG
-from instruments.python.uncabled.phsen.request_phsen import phsen_datalogger
+    update_dataset
+from instruments.python.uncabled.request_ctdbp import ctdbp_datalogger
+
+CONFIG = yaml.safe_load(open('instruments\\python\\config.yaml'))
 
 
 def main():
@@ -13,12 +16,12 @@ def main():
     # Ocean Observatories Initiative web site. The last two will set path and naming conventions to save the data
     # to the local disk
     site = 'CE02SHSM'           # OOI Net site designator
-    node = 'RID26'              # OOI Net node designator
-    sensor = '06-PHSEND000'     # OOI Net sensor designator
-    stream = 'phsen_abcdef_dcl_instrument'  # OOI Net stream name
+    node = 'RID27'              # OOI Net node designator
+    sensor = '03-CTDBPC000'     # OOI Net sensor designator
+    stream = 'ctdbp_cdef_dcl_instrument'  # OOI Net stream name
     method = 'telemetered'      # OOI Net data delivery method
     level = 'nsif'              # local directory name, level below site
-    instrmt = 'phsen'           # local directory name, instrument below level
+    instrmt = 'ctdbp'           # local directory name, instrument below level
 
     # We are after telemetered data. Determine list of deployments and use the last, presumably currently active,
     # deployment to determine the start and end dates for our request.
@@ -29,15 +32,15 @@ def main():
 
     # request and download the data
     r = m2m_request(site, node, sensor, method, stream, start, stop)
-    phsen = m2m_collect(r, '.*PHSEN.*\\.nc$')
-    phsen = phsen.where(phsen.deployment == deploy, drop=True)  # limit to the deployment of interest
+    ctdbp = m2m_collect(r, '.*ctdbp.*\\.nc$')
+    ctdbp = ctdbp.where(ctdbp.deployment == deploy, drop=True)  # limit to the deployment of interest
 
     # clean-up and reorganize
-    phsen = phsen_datalogger(phsen)
-    phsen = update_dataset(phsen, vocab['maxdepth'])
+    ctdbp = ctdbp_datalogger(ctdbp, burst=False)
+    ctdbp = update_dataset(ctdbp, vocab['maxdepth'])
 
     # save the data
-    out_path = os.path.join(CONFIG['base_dir']['m2m_base'], site.lower(), level, instrmt)
+    out_path = CONFIG['base_dir']['m2m_base'] + ('\\%s\\%s\\%s' % (site.lower(), level, instrmt))
     out_path = os.path.abspath(out_path)
     if not os.path.exists(out_path):
         os.makedirs(out_path)
@@ -45,7 +48,7 @@ def main():
     out_file = ('%s.%s.%s.deploy%02d.%s.%s.nc' % (site.lower(), level, instrmt, deploy, method, stream))
     nc_out = os.path.join(out_path, out_file)
 
-    phsen.to_netcdf(nc_out, mode='w', format='NETCDF4', engine='netcdf4')
+    ctdbp.to_netcdf(nc_out, mode='w', format='NETCDF4', engine='netcdf4')
 
 
 if __name__ == '__main__':
