+++ instruments/python/uncabled/examples/phsen/telemetered_phsen.py
@@ -4,35 +4,36 @@ import os
 
 from instruments.python.common import list_deployments, deployment_dates, get_vocabulary, m2m_request, m2m_collect, \
     update_dataset, CONFIG
-from instruments.python.uncabled.phsen.request_phsen import phsen_instrument
+from instruments.python.uncabled.request_phsen import phsen_datalogger
 
 
 def main():
     # Setup needed parameters for the request, the user would need to vary these to suit their own needs and
-    # sites/instruments of interest. Site, node, sensor and stream names can be obtained from the Ocean Observatories
-    # Initiative web site
+    # sites/instruments of interest. Site, node, sensor, stream and delivery method names can be obtained from the
+    # Ocean Observatories Initiative web site. The last two will set path and naming conventions to save the data
+    # to the local disk
     site = 'CE02SHSM'           # OOI Net site designator
     node = 'RID26'              # OOI Net node designator
     sensor = '06-PHSEND000'     # OOI Net sensor designator
-    stream = 'phsen_abcdef_instrument'  # OOI Net stream name
-    method = 'recovered_inst'   # OOI Net data delivery method
+    stream = 'phsen_abcdef_dcl_instrument'  # OOI Net stream name
+    method = 'telemetered'      # OOI Net data delivery method
     level = 'nsif'              # local directory name, level below site
     instrmt = 'phsen'           # local directory name, instrument below level
 
-    # We are after recovered instrument data. Determine list of deployments and use a recent one to determine
-    # the start and end dates for our request.
+    # We are after telemetered data. Determine list of deployments and use the last, presumably currently active,
+    # deployment to determine the start and end dates for our request.
     vocab = get_vocabulary(site, node, sensor)[0]
     deployments = list_deployments(site, node, sensor)
-    deploy = deployments[-3]
+    deploy = deployments[-1]
     start, stop = deployment_dates(site, node, sensor, deploy)
 
     # request and download the data
     r = m2m_request(site, node, sensor, method, stream, start, stop)
-    phsen = m2m_collect(r, '.*phsen.*\\.nc$')
+    phsen = m2m_collect(r, '.*PHSEN.*\\.nc$')
     phsen = phsen.where(phsen.deployment == deploy, drop=True)  # limit to the deployment of interest
 
     # clean-up and reorganize
-    phsen = phsen_instrument(phsen)
+    phsen = phsen_datalogger(phsen)
     phsen = update_dataset(phsen, vocab['maxdepth'])
 
     # save the data
