+++ instruments/python/utilities/harvesters/harvest_cp_phsen.sh
@@ -12,7 +12,7 @@
 
 # set the base directory python command for all subsequent processing
 cd /home/ooiuser/code/ooi-data-explorations
-PYTHON="python -m instruments.python.uncabled.phsen.request_phsen"
+PYTHON="python -m instruments.python.uncabled.request_phsen"
 
 ### CP01CNSM ###
 BASE_FLAGS="-s CP01CNSM -n RID26 -sn 06-PHSEND000"
